+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.36.1] - 2025-06-04
+
+### Changed
+
+- Changed the `runner` method of the `Client` to take an optional `cluster` argument. If not provided, the default cluster will be used.
+- Changed the `create_cron_automation` and `create_storage_event_automation` methods of the `AutomationClient` to take an optional `cluster` argument. If not provided, the default cluster will be used.
+- Changed the `submit` method of the `JobClient` to take an optional `cluster` argument. If not provided, the default cluster will be used.
+
 ## [0.36.0] - 2025-05-27
 
 ## Added
@@ -167,7 +175,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Released packages: `tilebox-datasets`, `tilebox-workflows`, `tilebox-storage`, `tilebox-grpc`
 
 
-[Unreleased]: https://github.com/tilebox/tilebox-python/compare/v0.36.0...HEAD
+[Unreleased]: https://github.com/tilebox/tilebox-python/compare/v0.36.1...HEAD
+[0.36.1]: https://github.com/tilebox/tilebox-python/compare/v0.36.0...v0.36.1
 [0.36.0]: https://github.com/tilebox/tilebox-python/compare/v0.35.0...v0.36.0
 [0.35.0]: https://github.com/tilebox/tilebox-python/compare/v0.34.0...v0.35.0
 [0.34.0]: https://github.com/tilebox/tilebox-python/compare/v0.33.1...v0.34.0